#39 Make controltag cache factories more permissive

+++ c2mon-server/c2mon-server-configuration/src/test/java/cern/c2mon/server/configuration/api/EverythingEverConfigurationTest.java
@@ -477,7 +477,7 @@ public class EverythingEverConfigurationTest extends ConfigurationCacheTest {
     cern.c2mon.server.common.process.Process processObj = processCache.get(expectedObject.getProcessId());
     assertTrue(processObj.getEquipmentIds().contains(expectedObject.getId()));
     assertNotNull(commFaultTagCache.get(expectedObject.getCommFaultTagId()));
-    assertEquals(expectedObject.getId(), commFaultTagCache.get(cacheObject.getCommFaultTagId()).getSupervisedId());
+    assertEquals(expectedObject.getId(), (long) commFaultTagCache.get(cacheObject.getCommFaultTagId()).getSupervisedId());
     assertNotNull(equipmentMapper.getItem(10L));
 
     verify(communicationManager);
@@ -629,7 +629,7 @@ public class EverythingEverConfigurationTest extends ConfigurationCacheTest {
     cern.c2mon.server.common.equipment.Equipment equip = equipmentCache.get(expectedObject.getParentId());
     assertTrue(equip.getSubEquipmentIds().contains(expectedObject.getId()));
     assertNotNull(commFaultTagCache.get(expectedObject.getCommFaultTagId()));
-    assertEquals(expectedObject.getId(), commFaultTagCache.get(cacheObject.getCommFaultTagId()).getSupervisedId());
+    assertEquals(expectedObject.getId(), (long) commFaultTagCache.get(cacheObject.getCommFaultTagId()).getSupervisedId());
     assertNotNull(subEquipmentMapper.getItem(20L));
 
     verify(communicationManager);
